Changed the type from 8 bit to 16 bit

+++ src/fakeLightSensor.cpp
@@ -6,8 +6,8 @@ FakeLightSensor::FakeLightSensor()
     srand(time(NULL));
 }
 
-uint8_t FakeLightSensor::lightLevel()
+uint16_t FakeLightSensor::lightLevel()
 {
-    uint8_t random = rand() % 1000 + 1;
+    uint16_t random = rand() % 1000 + 1;
     return random;
 }
